requirements.txt
matplotlib==3.7.5
openai==1.23.3
pandas==2.0.3
python-dotenv==1.0.1
seaborn==0.13.2
sentence-transformers==2.7.0
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2